+++ lib/ts/axios.ts
@@ -130,7 +130,7 @@ export function responseInterceptor(axiosInstance: any) {
 export default class AuthHttpRequest {
     private static refreshTokenUrl: string | undefined;
     static websiteRootDomain: string;
-    static sessionExpiredStatusCode = 440;
+    static sessionExpiredStatusCode = 401;
     static initCalled = false;
     static apiDomain = "";
     private static refreshAPICustomHeaders: any;
@@ -141,11 +141,17 @@ export default class AuthHttpRequest {
 
     static init(
         refreshTokenUrl: string,
-        sessionExpiredStatusCode?: number,
         websiteRootDomain?: string,
-        refreshAPICustomHeaders?: any
+        refreshAPICustomHeaders?: any,
+        sessionExpiredStatusCode?: number
     ) {
-        FetchAuthRequest.init(refreshTokenUrl, sessionExpiredStatusCode);
+        FetchAuthRequest.init(
+            refreshTokenUrl,
+            undefined,
+            websiteRootDomain,
+            refreshAPICustomHeaders,
+            sessionExpiredStatusCode
+        );
         AuthHttpRequest.refreshTokenUrl = refreshTokenUrl;
         AuthHttpRequest.refreshAPICustomHeaders = refreshAPICustomHeaders === undefined ? {} : refreshAPICustomHeaders;
         AuthHttpRequest.websiteRootDomain =