+++ lib/ts/index.ts
@@ -118,7 +118,7 @@ export function getDomainFromUrl(url: string): string {
  */
 export default class AuthHttpRequest {
     private static refreshTokenUrl: string | undefined;
-    private static sessionExpiredStatusCode = 440;
+    private static sessionExpiredStatusCode = 401;
     private static initCalled = false;
     static originalFetch: any;
     private static apiDomain = "";
@@ -142,10 +142,10 @@ export default class AuthHttpRequest {
 
     static init(
         refreshTokenUrl: string,
-        sessionExpiredStatusCode?: number,
         viaInterceptor?: boolean,
         websiteRootDomain?: string,
-        refreshAPICustomHeaders?: any
+        refreshAPICustomHeaders?: any,
+        sessionExpiredStatusCode?: number
     ) {
         if (viaInterceptor === undefined) {
             if (AuthHttpRequest.viaInterceptor === undefined) {