+++ test/server/index.js
@@ -199,12 +199,12 @@ app.use(
         onTryRefreshToken: (err, req, res) => {
             res.header("Access-Control-Allow-Origin", "http://127.0.0.1:8080");
             res.header("Access-Control-Allow-Credentials", true);
-            res.status(440).send();
+            res.status(401).send();
         },
         onUnauthorised: (err, req, res) => {
             res.header("Access-Control-Allow-Origin", "http://127.0.0.1:8080");
             res.header("Access-Control-Allow-Credentials", true);
-            res.status(440).send();
+            res.status(401).send();
         }
     })
 );