+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.2.0] - 2020-08-11
+### Changed
+- Changed the default session expiry status code to 401
+- Changed function signature of `init` for `axios` and `fetch`
+
 ## [4.1.5] - 2020-07-30
 ### Added
 - Function to get Refresh URL's domain